rolling_update: enforce default value for client_update_batch
`ansible_forks` isn't something always set by ansible.

This commit allows to either pass `client_update_batch` as an extra variable
that will be the *number* of node per batch or
`client_update_batch_percentage` as an extra variable as well that will
be the percentage of the total number of client node in a decimal
representation (eg: `0.5` for `50%`).
If none of them are passed, the default value is 0.2 (20%).

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1650184

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>