Ability to tune kernel parameters
ceph-ansible is already capable of tuning kernel parameters but only
parameters that can change on the live system.
This commit allows us to change specific kernel parameters that require
a reboot.
So during the initial phase, we play the ceph-infra role which will
apply some tuning if specified. Then all the machines will be rebooted.
We wait up to 15min for that reboot before failing, assuming the machine
is stuck in its init sequence.

To change a specific parameter just set:

tune_kernel_parameters: "key=value key=value"

This implementation has some assumptions:

* the system is RHEL8
* the 'grubby' command does things correctly

Things we don't do:

* validate the kernel param is valid
* validate that has been set (we trust 'grubby' for this)

We only print the bootloader cmdline so you check if the system has the
correct value.

Signed-off-by: Sébastien Han <seb@redhat.com>