Updated rbd_client_admin_socket_path to match ceph-common role
Updated the owner, group and mode for rbd_client_admin_socket_path in order to match how rbd_client_admin_socket_path is defined in /roles/ceph-common/tasks/create_rbd_client_dir.yml. The default for the mode in the ceph-common role as defined by rbd_client_directory_mode is 0770 while the mode is hard coded as 0755 in the ceph-rgw role. This was causing a mode clash on /var/run/ceph when the ceph-rgw role is enabled.

Signed-off-by: Kevin Coakley <kcoakley@sdsc.edu>