Adjust openstack_keys
Added missing class-read object_prefix rbd_children to glance
so that they can unprotect images. As these permissions should
be set globally on cluster but not only on the specific pool,
we need to add them explicitly.

Also since glance do not need write access to volumes pool and
set permission is static (do not rely on openstack_cinder_pool)
we can easily drop it.

Signed-off-by: Dmitriy Rabotyagov <drabotyagov@vexxhost.com>